feat: add client-side pagination to queues list
Extract shared PaginationControls component from duplicated alerts
pagination code and reuse it for the new queues table pagination.
Page resets on filter change and page size change.
